pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["CuasarN1 <55405975+CuasarN1@users.noreply.github.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
pydantic-settings = "^2.5.2"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"